+++ test/files/pos/t4365/a_1.scala
@@ -2,8 +2,8 @@ import scala.collection._
 
 trait SeqViewLike[+A,
                   +Coll,
-                  +This <: SeqView[A, Coll] with SeqViewLike[A, Coll, This]]
-  extends Seq[A]   with GenSeqViewLike[A, Coll, This]
+                  +This <: SeqView[A, Coll] with SeqViewLike[A, Coll, Nothing]]
+  extends Seq[A]   with GenSeqViewLike[A, Coll, Nothing]
 {
 
   trait Transformed[+B] extends super[GenSeqViewLike].Transformed[B]
